Fixes Two Broken Specs
The Spree::Avatax::SalesShared spec was broken, because it created a tax
adjustment with no source, which I believe to be impossible. Using the
`tax_adjustment` factory for the spec setup resolves the error.

The Spree::Avatax::ShortShipReturnInvoice is more troubling, because I don't
know why changing the magic number "fixed" the spec. I don't understand why the
outcome was different. This may bear more investigation.