Complexify Taxable Orders.
Taxable orders is somewhat of a large item for GV2. Avalara works
wonderful when dealing with North American or USA and CA orders.
However, Solidus itself can get to the bottom of VAT oriented taxes in a
much more elegant, and 'at the application level' manner.

Make sure we only create an invoice and call out to the Avatax service if
the order is under the CA and US stores. This also includes a slight
refactoring to tuck knowledge of a POS order into this taxable_order
method.