Integrate Avatax More Cleanly Into Solidus
Until now, Avatax would generate a sales-invoice when an order transitioned to
the payment state. This occurred irrespective of the tax-rates applicable to
the order, and necessitated additional logic to determine whether the order was
taxable and whether the order was Avatax-taxable.

This commit alters the extension's behaviour to act, well, like any other
Solidus tax-rate, via the `Spree::TaxRate#adjust` method. As a result, no calls
are made to Avalara unless the order belongs to a zone with an
Avatax-calculated tax-rate.