Only Commit If Avatax Adjustments Exist
The callback triggered by an order reaching the 'complete' state would attempt
to commit the Avalara sales invoice, whether or not one existed, or _should_
exist. For example, an order not covered by Avatax would expect not to have an
Avalara sales invoice. The effect, however, was to trigger a commit attempt,
which would raise an exception.

This commit makes the commit attempt conditional on the presence of an Avatax
tax adjustment. This will handle all cases where it is expected to not have a
sales-invoice: where other tax-rates are used, or jurisdictions in which the
business does not need to collect taxes.