Reduces Unnecessary Duplicated Sales Invoice Generation
An Avatax sales invoice doesn't need to be generated each time tax-adjustment
is calculated for a single line-item or shipment.  Once per order is just fine,
thankyouverymuch. This change in behaviour is achieved by overriding
`Spree::Tax::OrderAdjuster#adjust!`.  When all the tax-rates for the order's
zone are Avatax-calculated, a sales invoice is generated, and the rest of the
operation is skipped. The normal adjustment implementation is followed in any other case.

Additionally, a sales-invoice is no longer generated in Avatax-calculated tax
rates. Calling `TaxRate#adjust` in these cases in a no-op.