(FC) Change customer code from email to order number (#15)
* Make customer code configurable

This change is motivated by a hard to work around flow during ApplePay
checkout flow where we only get the customer's information like their
email at the very end step.

* Fix a failing spec

Validation failed: Payment method can't be blank was getting thrown for
these 2 tests, and yes checking the payment class source file correctly
points out that this is required. Invoking the payment factory is a nice
solution to this.

* Remove need for an error

We wrap the value into a proc as to make it conform to what the gem
expects internally.